Replace Python sqlite3 module by apsw.
http://rogerbinns.github.io/apsw/

The Python sqlite3 module is built-in to the CPython distribution and
provides a DB-API interface to sqlite3.  However, its support for
many sqlite3-specific features is limited.

apsw is an external module that follows the sqlite3 C API closely and
supports everything you can do in C.  apsw also lacks miscellaneous
brain damage that typifies the Python sqlite3 module.

In particular, we will want to be able to write modules for virtual
tables in order to implement SIMULATE, which we can do with apsw but
not with the Python sqlite3 module.