Add BEGIN, ROLLBACK, COMMIT to BQL.
No savepoints yet -- more trouble than they're worth.

No options like in SQLite: the commands are just

    BEGIN
    ROLLBACK
    COMMIT

not BEGIN TRANSACTION, not START TRANSACTION, no transaction types,
&c.  Those may come later.

Negative side effect of this change: a little less happens inside a
transaction (really, inside a savepoint) in bdb.execute, so that
bdb.execute('BEGIN') can do anything sensible.  So you may have to
wrap more in a savepoint yourself:

    with bdb.savepoint():
        results = list(bdb.execute('...'))

Positive side effect of this change: it will be possible to deal with
transactions using only the BQL language, not using the Python API
around it.  So you could write a thin RPC gizmo between, say, R and
Python, to exchange BQL phrases, and transactions would work.

In particular, if you want to get a large table of output involving
evaluating many model-dependent functions, the caller can start with
BEGIN, then execute the query, and then execute COMMIT (or ROLLBACK),
and the effect will be similar to the above fragment of Python code
wrapped in `with bdb.savepoint():'.