Create a bayeslite-specific engine API and move crosscat to it.
Use this to make metadata creation specific to the metamodel.  Soon,
most of bqlfn.py will move in there.

User-visible effect is that instead of writing

    cc = crosscat.LocalEngine.LocalEngine(seed=0)
    bayesdb_register_metamodel(bdb, 'crosscat', cc)

it will be necessary to write

    import bayeslite.crosscat

    cc = crosscat.LocalEngine.LocalEngine(seed=0)
    engine = bayeslite.crosscat.CrosscatEngine(cc)
    bayesdb_register_metamodel(bdb, 'crosscat', engine)

with the caveat that this API is likely to change again soon because
I don't like the overuse of the word `engine'.  (Probably something
like bayeslite.croscat.CrosscatMetamodel(cc) instead.)