Flesh out ALTER TABLE and ALTER GENERATOR a little.
Support renaming tables and generators.  New code to partially
support renaming columns too, but it is not yet finished: I forgot
that we'll need to map them to the SQL names when compiling BQL,
which requires keeping track of lexical environments in the compiler,
which is the right thing to do but we don't do it yet.