Simplify 5->6 schema upgrade.
This was what the previous commit was supposed to be.  Fortunately,
the only difference between that and this change is that the previous
commit included an extra constraint, tabname != name, in
bayesdb_generator, which is not sufficient: we need the generator's
name to be distinct from *all* table names, not just all table names.
We rely on code in bayesdb to ensure that, because expressing the
full constraint in SQL is tricky.

So this change is safe, and simplifies the schema upgrade.