Implement INFER x AS xi CONFIDENCE xc.
This is a kind of evil kludge: we cram two values into one column as
a JSON blob, and then pick it apart in a post-processing query.  In
the process, we lose automatic column names for columns that are not
named explicitly and are not column references themselves.  Tough.

This adds the method infer_confidence to metamodels, and removes the
now-unnecessary value argument from infer.