Add subsampling in Crosscat.
Notation is SUBSAMPLE(1000) or SUBSAMPLE(OFF) in generator schema.

First 1000 rows will be fed to Crosscat.  BQL functions on other rows
will be ephemerally inserted into Crosscat each time to compute the
function.  Currently no way to permanently add rows to the subsample
or do analysis on rows not inserted.