Another bug: Each of the Geweke chains needs its own synthetic data.
Otherwise what I have is one sample from the Markov chain whose state
space consists of 500 samples of the latents, etc.  The effect on the
nig normal example is that, with a large data set, my 500 Geweke
chains concentrate on the best Gaussian approximation to the prior
predictive, which is not at all what I want.