Compute chi2_sf with fast-converging sequences, not with Monte Carlo.
Gets better results faster.  We now have low relative error, not low
absolute error, far out in the tail of the distribution, so update
the tests accordingly.

XXX Do the same for F and student_t.  Requires computing incomplete
beta integrals, which is a little hairier than computing incomplete
gamma integrals as chi^2 requires.  Need to find a good reference for
this.