Redo mechanism to autogenerate grammar.py when appropriate.
Save the SHA-256 hash of grammar.y and grammar.py in grammar.sha256,
and regenerate only if they don't match.

(Funny property: the test is symmetric with respect to inputs and
outputs, so we don't even need to mark each line for whether it's
input or output.  We could generalize this to any number of inputs or
outputs too.)