Don't ever let cursor.description be None.
This is nonsense inherited from the Python sqlite3 module.
Unfortunately, it is not straightforward to override this nonsense,
because the property is made read-only in C, so you can't do

    cursor.description = []

to make it sane.  So we instead make a kludgey subclass of
sqlite3.Cursor, in order to avoid meticulously manifesting all the
relevant functionality of cursors.