Symlink to /usr/bin/python2.7, not /usr/bin/python.
/usr/bin/python is a magic xcode binary that breaks if you try to
exec it with argv[0] set to python2.7 -- it appends 2.7 to argv[0]
and tries to execute the result inside some /System/Library lair, but
there is no python2.72.7 there so it fails.

Hard-coding a reference to 2.7 is OK here because we already do it
elsewhere, so it can't get much worse and it's still greppable.