Implement random subsampling in Crosscat metamodel.
Fixes Github issue #40.

The seed is currently a deterministic function of the total number of
rows and the number of subsampled rows.  In the future, perhaps we
ought to let the user specify a seed in case they want to explore
other possible subsamplings.  We might do this when we store in the
Crosscat state a random seed we need for Crosscat's purposes anyway,
rather than having it only in-memory in the process as it is now.