Updates to CrossCat deps to build osx dmg on mac mini.
First, all those deps were for Crosscat, not for BdbContrib. Mislabelled.

Next, we weren't doing anything with the libboost dir anymore, but now we do need to, and it's to set an environment variable.

We now look for one of the headers that we actually use.

Finally, no need to separately set up for bayeslite anymore, because we fixed its setup.py earlier.