Test that SIMULATE x GIVEN x = 40000 always yields 40000 in crosscat.
This is a very simple-minded test -- it should be generalized to a
generic quality test suite for any metamodel later.

For Github issue #175, Github issue crosscat#34.

Bump Crosscat requirement to 0.1.29, in which the Crosscat bug is
fixed.

(Yes, this test should be in Crosscat, not in bayeslite.  However,
the Crosscat automatic tests are a sorry mess, and putting the test
here is much more expedient.)