Use vectorized numpy routines for computing correlation.
Avoided this before because we were avoiding a dependency on numpy
before.  It is no longer worthwhile to do so, and calculating
correlations has become a bottleneck for some demonstrations --
which, while in line with our claim that dependence probability and
mutual information are better, does not really do us any favours by
needlessly slowing things down.

This changes the results very slightly, owing to the substitution of
numpy.sum, which does naive accumulated summation, for float_sum,
which does Kahan-Babuska compensated summation.