Omit bdb.prng for now until we are happier with the API.
I put `weak' in the name of all the methods of the WeakPRNG class, of
which this is an instance, to emphasize that it does not provide
backtracking resistance and is consequently unfit for cryptographic
key generation.

So perhaps bdb.prng should be called bdb.weakprng, although then
users might think that bdb.weakprng gives lower-quality output than
bdb.py_prng and bdb.np_prng, which is backwards.

Or perhaps one might argue that this is not likely to be used for
crypto purposes, and the `weak' lexeme should be omitted from all its
methods so you can say bdb.prng.random32() &c.