Make schema upgrades explicit.
New public API function:

    bayesdb_upgrade_schema(bdb)

Upgrades bdb to the latest schema version.

As an undocumented internal feature, you can also specify version=42
for a specific schema version.  But specific schema versions need not
be user-facing for now.

Use this to fail noisily on attempts to use sessions with schema
version 6, which was produced by bayeslite 0.1 and 0.1.1 and lacks
the tables necessary for sessions.

In this case, adding the relevant tables for sessions would not break
bayeslite 0.1 and 0.1.1 -- but there's not (yet) a mechanism for
describing the oldest schema version that the current one is
compatible with, and this helps to exercise the problem of schema
upgrades.  (In the past, in prereleases, we have done such
incompatible schema changes.)

Later, we can decide to desupport older schema versions.