Implement simulation from the joint distribution.
The strategy is to create a temporary ripl for each simulation
drawing randomly from the 1 ... modelno. Then we perform a sequence
of observes for the constraints, followed by 100 iterations of mh
only on the row latent scope and rowid block, followed by a sequence
of predict, then further inference per-predict (to ensure samples are
from the approximate joint distribution).