Don't hang onto this cursor longer than needed.
Doing so causes apsw to fail with a `database is locked' error.  It
is unclear to me why switching from the Python sqlite3 module to apsw
caused the symptom to arise -- but it likely indicates a bug in the
Python sqlite3 module whereby cursors are stepped one row too fast,
so that the Python sqlite3 module can discover that the iteration is
complete sooner than apsw does.  Hence this simplification likely
does not merely paper over a bug in apsw.