Correct calculation of Pearson r^2 correlation p-values.
The Monte Carlo evaluation of the Student's t CDF is so inaccurate
that we get only one correct digit in this case, so for now, we just
require only one correct digit in the automatic tests.  Fixing this
requires either taking many more Monte Carlo samples, or numerical
analysis to evaluate incomplete beta ratios, which are notoriously
hairy to evaluate.

(The difficulty of their evaluation is, I suspect, the reason why one
Mr Karl Pearson, F.R.S., decided to invent the approximate chi^2 test
back in 1900 or so instead of evaluating a binomial CDF exactly for
Weldon's dice data!)