Revert `model schema' concept.  Introduce `population' instead.
This *concept* that I provisionally named `model schema' is already
essentially what we called a `generator'.  The real change we mean to
apply is to refocus most of BQL on populations instead of generators
(and maybe later we will rename `generator' to `metamodel').

Simply introducing populations now will be easier than introducing a
concept that is functionally equivalent to one we already have but
not proprely hooked into anything.