Tweak order of evaluating likelihood vs pdf.
Shouldn't affect results but one test checks what was most recently
passed through to the metamodel's logpdf_joint method, which changes
if we do likelihood last.

XXX Terminology is wrong?  Likelihood is P(Q|C) and prior is P(C),
but I called P(C) the `likelihood' of a generator and P(Q|C) the pdf.