New table bayesdb_latent for names of latent variables.
Names are otherwise stored in bayesdb_column.

XXX In retrospect, perhaps this separation of table was a bad idea,
and perhaps we ought to have keyed everything on the variable name --
which may make renaming them more painful, but probably less badly so
than implementing variable renaming would be right now anyway (by
teaching the compiler to track lexical scope).