Associate latent variables with generators.
Add bayesdb_add_latent to add a latent variable from a metamodel's
create_generator operation.

Two generators may not currently share an overlapping latent variable
name.  (XXX It is not clear that this is the right design choice.  We
could require the variable to be qualified by the generator's name if
ambiguous.  But we'll defer that for now.)