Fix #468.
In particular, rather than just pass in distargs['cctypes'] and
distargs['ccargs'] which apply to the cgpm inputs, we instead pass in
distargs['outputs']['stattypes/statargs'] and similarly for inputs. This
information is required to implement the CGPM interface correctly, and
is required for the pca/kde/knn cgpms.

The interface of State/Engine has NOT been updated (it only applies to
new-style cgpms which implement the default model). A ticket is open to
chagne the interface of State/Engine to match these changes.