Allow INFER EXPLICIT PREDICT col [AS <colname>] [CONFIDENCE <confname>]
In the past, the phrase CONFIDENCE <confname> was obligatory. Now, if
a user does not request CONFIDENCE then the returned table shall not
include that column, for example:

CREATE TABLE flower_pc_components AS
    INFER EXPLICIT
        name,
        PREDICT flower_pc1 CONFIDENCE _conf1 USING 10 SAMPLES,
        PREDICT flower_pc2 CONFIDENCE _conf2 USING 10 SAMPLES
    FROM iris_p
    MODELED BY iris_g;

can now be written as

CREATE TABLE flower_pc_components AS
    INFER EXPLICIT
        name,
        PREDICT flower_pc1 USING 10 SAMPLES,
        PREDICT flower_pc2 USING 10 SAMPLES
    FROM iris_p
    MODELED BY iris_g;

the latter returns a table with just three columns, as oppose to the
five returned by the former.