Implement retrieval of constraints from base table given rowid.
If the user specifies a `rowid` in `CGPM_Metamodel.simulate` which exists
in the base table but has not been incorporated into the cgpm (i.e.
by inserations after `INITIALIZE`), then load any values in the table
and add them to the `constraints`.

An error is returned in the user attempts to specify `constraints` which
contradict actual values in the table. Only values which are `None` or
SQL `NULL` may be overriden.

Method `predict_confidence` is updated to specify the `rowid` rather
than manually populate constraints.