Push bql_compiler through to subqueries.
New IBQLCompiler method returns the implicit target variable, except
I mixed things up with of/with so that it is confusingly used as a
reference variable, which doesn't matter because depprob and MI are
symmetric anyway, but we should sort it out.

XXX Should rename `bql_compiler' to `environment' and simplify.