Remove unused tabname from bayesdb_generator.
Ideally we would specify

  FOREIGN KEY(population_id, tabname) REFERENCES
    bayesdb_population (id, tabname)

However the above would require (id, tabname) to be the primary key
of bayesdb_population, which we cannot do since bayesdb_population.id
is AUTOINCREMENT and therefore must be the only primary key in the table.
This commit causes 1 additional query to occur, but reduces the chance of
malforming the bayesdb_generator table with an invalid (population_id, tabname)
pair that we cannot use SQL to enforce actually corresponds to a valid pair.