Initial commit for bqlmath module (intentional bug, for versioning).
For some strange reason, invoking db.createscalarfunction in the loop
causes all the sqlite math funcs to refer to a single one of the math.*
functions. pylint complains "cell variable fn defined in loop" which
suggests a candidate reason, but I cannot seem to figure it out.

Writing the loop functionally, by embedding the body of the loop in
a helper function e.g. install_func solves the issue. Should revisit
in the future.