Update tests/test_vscgpm to match github.com:probcomp/cgpm@20180116-fsaad-vscgpm-inputs-dicts
Main issue with getting this to work is that logpdf is now correctly
implemented, but in the Venture program the output simulate_period is
deterministic since '+' is of type deterministic SP. Therefore, we need
to rewrite the Venture program, which is not something high priority.

Therefore, merge of 20160116-fsaad-vscgpm-inputs-dicts into cgpm master
shall be delayed until further notice.