Fix use of global rand state.
The code uses bdb.np_prng instead of np.random everywhere now. Further,
The seed is decorated with the @stochastic decorator. Because of the
way this decorator works, I can use pytest fixtures with it anymore.
Thus, the `backend` fixture was removed (`backend=cgmp` was commented
out anyways because it was too slow.